+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Format amounts using localized decimal and group separator
 - Support pasting different localized number formats, i.e. dot and comma separated amounts
 - Fix BitBoxApp crash on GrapheneOS and other phones without Google Play Services when scanning QR codes.
+- Add DMG installer for macOS
 
 ## 4.42.0
 - Preselect backup when there's only one backup available